+++ tests/batch_cosign_e2e.rs
@@ -20,7 +20,7 @@ use rand010::{SeedableRng, rngs::StdRng};
 
 const LEVEL: RegevSecurityLevel = RegevSecurityLevel::Test;
 
-fn member_info(slot: u8, keys: &MemberKeys) -> MemberInfo {
+fn member_info(slot: u16, keys: &MemberKeys) -> MemberInfo {
     MemberInfo {
         slot,
         pk_g: keys.pk_g(),
